Creation of Gabor filter banks
Initial work on the code to create whole Gabor filter banks from the
combination of sets of parameters. These parameters sets are:

- A set of bandwidths
- A set of orientations
- A set of spatial frequencies

The number of filters in the generated bank equal the number of
possible combinations of the elements of each set (picking 1 from each)

NOTE this is a partial commit, the code is not finished yet.

(addresses #2)